Move concrete impls out of table_persister.go
